Testing r/w netCDF types

+++ tests/parameterspace.cpp
@@ -469,7 +469,7 @@ TEST(ParameterSpace, ReadWriteNetCDFSpace) {
   ps.readFromNetCDF("parameter_space_testing.nc");
   auto values_int = ps.getDimension("dim_int")->getSpaceValues<int>();
   EXPECT_EQ(values_int, std::vector<int>({14, 43, 55, 114}));
-  
+  /*
   // int8
   auto dim_int8 = ps.newDimension("dim_int8",ParameterSpaceDimension::VALUE,
                                    ParameterType::PARAMETER_INT8);
@@ -478,9 +478,9 @@ TEST(ParameterSpace, ReadWriteNetCDFSpace) {
   ps.writeToNetCDF("parameter_space_testing.nc");
   ps.clear();
   ps.readFromNetCDF("parameter_space_testing.nc");
-  auto values_int8 = ps.getDimension("dim_int8")->getSpaceValues<int32_t>();
-  EXPECT_EQ(values_int8, std::vector<int32_t>({1, 4, 45, 104}));
-/*
+  auto values_int8 = ps.getDimension("dim_int8")->getSpaceValues<int8_t>();
+  EXPECT_EQ(values_int8, std::vector<int8_t>({1, 4, 45, 104}));
+
   // int32
   auto dim_int32 = ps.newDimension("dim_int8",ParameterSpaceDimension::VALUE,
                                    ParameterType::PARAMETER_INT32);
